+++ src/test/java/UItests/contacts/ContactCreate.java
@@ -8,8 +8,8 @@ public class ContactCreate extends BaseTest {
 
     @Test
     public void testCreateNewContact() {
-        openCreateNewContact();
-        fillInNewContactFields(new newContactData("FirstName", "LastName", "NickName", "+380009999561", "any_valid_test_email@yopmail.com"));
-        clickSubmitBtn();
+        appmngr.openCreateNewContact();
+        appmngr.fillInNewContactFields(new newContactData("FirstName", "LastName", "NickName", "+380009999561", "any_valid_test_email@yopmail.com"));
+        appmngr.clickSubmitBtn();
     }
 }
